pyproject.toml
[project]
name = "organia-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "peewee>=3.17.6",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.9",
    "torch>=2.4.1",
    "transformers>=4.45.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]
